requirements.txt
backoff
faiss-cpu
huggingface-hub
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchainhub
langsmith
sentencepiece
transformers
openai
python-dotenv
PyYAML
requests
scikit-learn
semantic-version
sentence-transformers
sentencepiece
tokenizers
torch
protobuf
pandas
flask-socketio